# openapi.yaml
openapi: 3.0.0
info:
  title: Finance Backend API
  version: 1.0.0
  description: API documentation for the Finance Backend
servers:
  - url: http://localhost:3000
paths:
  /users/login:
    post:
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /users:
    get:
      summary: List users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /users/{id}:
    put:
      summary: Update user details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    patch:
      summary: Partially update user details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /records:
    get:
      summary: Get records with optional filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: List of filtered records
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/Record'
                  total:
                    type: integer
    post:
      summary: Create a record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
      responses:
        '201':
          description: Record created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /records/{id}:
    put:
      summary: Update a record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
      responses:
        '200':
          description: Record updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete a record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Record deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /dashboard/summary:
    get:
      summary: Get overall dashboard summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalIncome:
                    type: number
                  totalExpense:
                    type: number
                  netBalance:
                    type: number
                  categoryTotals:
                    type: array
                    items:
                      type: object
                  recent:
                    type: array
                    items:
                      $ref: '#/components/schemas/Record'
                  monthly:
                    type: array
                    items:
                      type: object
  /dashboard/category-totals:
    get:
      summary: Get category totals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Category totals
          content:
            application/json:
              schema:
                type: object
                properties:
                  categoryTotals:
                    type: array
                    items:
                      type: object
  /dashboard/monthly-trends:
    get:
      summary: Get monthly trend metrics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Monthly trends
          content:
            application/json:
              schema:
                type: object
                properties:
                  monthlyTrends:
                    type: array
                    items:
                      type: object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - role
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [ADMIN, ANALYST, VIEWER]
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [ADMIN, ANALYST, VIEWER]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        status:
          type: string
    CreateRecordRequest:
      type: object
      required:
        - amount
        - type
        - category
        - date
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [INCOME, EXPENSE]
        category:
          type: string
        date:
          type: string
          format: date-time
        notes:
          type: string
    Record:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        type:
          type: string
        category:
          type: string
        date:
          type: string
        notes:
          type: string
        deleted:
          type: integer